slash and slide integration
