+++ Assets/Prefabs/Environment/SlashableWall.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 6313d634746a1ca4d88ea49a0448ad66
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
