+++ Assets/Prefabs/Environment/StabableCube.prefab.meta
