--- Assets/Scripts/Sword.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 23d4fef0cca1d704bb252cb88b6337eb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
